requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
h11==0.14.0
huggingface-hub==0.24.6
idna==3.10
langchain==0.3.22
langchain-community==0.3.20
langchain-core==0.2.11
langchain-groq==0.3.2
langchain-text-splitters==0.3.7
groq==0.6.0
pydantic==2.11.2
pydantic-core==2.33.1
python-dotenv==1.0.1
python-multipart==0.0.9
sentence-transformers==2.6.0
sniffio==1.3.1
starlette==0.46.1
transformers==4.42.4
typing-extensions==4.13.0
typing-inspection==0.4.0
uvicorn==0.34.0
langgraph==0.0.15
psycopg2-binary==2.9.9
python-jose==3.3.0
SpeechRecognition==3.10.1
pydub==0.25.1
requests==2.31.0
python-dateutil==2.8.2
langchain-huggingface==0.1.2
